refactor: 동아리 좋아요 관련 서비스 단위 테스트 분리
